refactor: #99 refactoring archiving apis
