+++ src/components/cards/popularArchivingCard/PopularArchivingCard.tsx
@@ -2,17 +2,16 @@ import React, { useState } from 'react'
 
 import { useNavigation } from '@react-navigation/native'
 import { ImageURISource } from 'react-native'
-import Config from 'react-native-config'
 import { useMutation, useQueryClient } from 'react-query'
 import { useRecoilValue } from 'recoil'
 
-import { patchScrapArchiving } from '@/apis/archiving'
+import { patchScrapArchiving } from '@/apis/archiving/Archiving'
 import { defaultIcons, defaultImages } from '@/assets'
 import PhotoIcon from '@/assets/icons/photo.svg'
 import ScrapIcon from '@/assets/icons/scrap.svg'
 import ScrapFillIcon from '@/assets/icons/scrap_fill.svg'
 import ScrapSmallIcon from '@/assets/icons/scrap_small.svg'
-import { ArchivingListContent } from '@/models/Archiving'
+import { ArchivingInfo } from '@/models/Archiving'
 import { MainNavigationProp } from '@/navigations/MainNavigator'
 import { CommunityCategoryState } from '@/state/CategoryState'
 import { colors } from '@/styles/colors'
@@ -29,7 +28,7 @@ import {
 } from './PopularArchivingCard.style'
 
 interface PopularArchivingCardProps {
-  item: ArchivingListContent
+  item: ArchivingInfo
 }
 
 /**
