+++ src/screens/contentList/ContentList.style.ts
@@ -1,11 +1,11 @@
 import styled, { css } from '@emotion/native'
 import { FlatList } from 'react-native'
 
-import { SimpleContent } from '@/models/SimpleContent'
+import { ContentCardInfo } from '@/models/ContentCard'
 import { colors } from '@/styles/colors'
 import { fonts } from '@/styles/fonts'
 
-export const ContentListContainer = styled(FlatList<SimpleContent>)``
+export const ContentListContainer = styled(FlatList<ContentCardInfo>)``
 
 export const ScrollContainer = styled.ScrollView`
   padding: 0px 20px;
