+++ src/screens/main/home/Home.tsx
@@ -6,7 +6,7 @@ import { ImageURISource, ListRenderItem, TouchableOpacity } from 'react-native'
 import { useInfiniteQuery, useQuery, useQueryClient } from 'react-query'
 import { useRecoilState, useRecoilValue } from 'recoil'
 
-import { getHomeArchivingList } from '@/apis/archiving'
+import { getHomeArchivingList } from '@/apis/archiving/ArchivingList'
 import { getUser } from '@/apis/user'
 import { defaultImages } from '@/assets'
 import SearchButton from '@/components/buttons/searchButton/SearchButton'
@@ -17,7 +17,7 @@ import EmptyItem from '@/components/emptyItem/EmptyItem'
 import { CategoryList } from '@/components/lists/categoryList/CategoryList'
 import { Loading } from '@/components/loading/Loading'
 import i18n from '@/locales'
-import { ArchivingListContent, MainArchivingListResponse } from '@/models/Archiving'
+import { ArchivingInfo, MainArchivingListResponse } from '@/models/Archiving'
 import { MainNavigationProp } from '@/navigations/MainNavigator'
 import { isCloseToBottom } from '@/services/InfiniteService'
 import { isWindowWidthSmallerThen } from '@/services/SizeService'
@@ -215,7 +215,7 @@ export const Home = () => {
 /**
  * renderItem
  */
-const renderItem: ListRenderItem<ArchivingListContent> = ({ item }) => {
+const renderItem: ListRenderItem<ArchivingInfo> = ({ item }) => {
   return (
     <ArchivingCard
       item={item}
