+++ src/screens/recycleBin/tabs/AllTab.tsx
@@ -1,14 +1,14 @@
 import React from 'react'
 
-import { ListRenderItem, ScrollView, TouchableOpacity, View } from 'react-native'
+import { ListRenderItem, ScrollView, TouchableOpacity } from 'react-native'
 import { useRecoilState } from 'recoil'
 
 import CheckIcon from '@/assets/icons/check_yellow.svg'
 import { ArchivingCard } from '@/components/cards/archivingCard/ArchivingCard'
 import ContentCard from '@/components/cards/contentCard/ContentCard'
 import i18n from '@/locales'
+import { ContentCardInfo } from '@/models/ContentCard'
 import { RecycleBinTabProps } from '@/models/Recycle'
-import { SimpleContent } from '@/models/SimpleContent'
 import { CheckArchivingState, CheckContentState } from '@/state/CheckState'
 
 import {
@@ -58,7 +58,7 @@ export const AllTab = ({ contents, archivings, editMode }: RecycleBinTabProps) =
   /**
    * ListRenderItem
    */
-  const renderItem: ListRenderItem<SimpleContent> = ({ item }) => {
+  const renderItem: ListRenderItem<ContentCardInfo> = ({ item }) => {
     return (
       <TouchableOpacity
         key={item.contentId}
@@ -66,6 +66,7 @@ export const AllTab = ({ contents, archivings, editMode }: RecycleBinTabProps) =
         disabled={!editMode}
       >
         <ContentCard
+          archivingId={0}
           contentId={item.contentId}
           contentTitle={item.contentTitle}
           contentType={item.contentType}
